requirements.txt
Django==3.2
psycopg2==2.8
python-dotenv==0.17
